cull at every second and importing the transactions
